% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{accordion_marker}
\alias{accordion_marker}
\title{accordion_marker: Cell Marker Accordion database}
\format{
A data table
}
\usage{
data(accordion_marker)
}
\description{
accordion_marker: Cell Marker Accordion database
}
