% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{bone_marrow_data}
\alias{bone_marrow_data}
\title{bone_marrow_data: Example dataset of human bone marrow from}
\format{
A Seurat object
}
\usage{
data(bone_marrow_data)
}
\description{
bone_marrow_data: Example dataset of human bone marrow from
}
