% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{cell_cycle_markers}
\alias{cell_cycle_markers}
\title{cell_cycle_markers: Collection of markers assiciated to cell cycle}
\format{
A data table
}
\usage{
data(cell_cycle_markers)
}
\description{
cell_cycle_markers: Collection of markers assiciated to cell cycle
}
