% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{cell_onto}
\alias{cell_onto}
\title{cell_onto: Cell ontology}
\format{
Cell ontology

matrix

A data table

A data table

A data table

A Seurat object

A Seurat object

Uberon ontology
}
\usage{
data(cell_onto)

data(counts)

data(disease_accordion_marker)

data(marker_table)

data(marker_table_pathway)

data(mouse_data)

data(retinal_data)

data(uberon_onto)
}
\description{
cell_onto: Cell ontology

counts: PBMC raw counts

disease_accordion_marker: Cell Marker Accordion disease database

marker_table: Table of markers associated to retinal cell types

marker_table_pathway: Table of markers associated to apoptosis

mouse_data: Mouse single-cell dataset

retinal_data: Retina single-cell dataset

uberon_onto: Uberon anatomy ontology
}
