% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{disease_accordion_marker}
\alias{disease_accordion_marker}
\title{disease_accordion_marker: Cell Marker Accordion disease database}
\format{
A data table
}
\usage{
data(disease_accordion_marker)
}
\description{
disease_accordion_marker: Cell Marker Accordion disease database
}
