% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{gene_description}
\alias{gene_description}
\title{gene_description: table containing gene description for human and mouse genes (human112_GRCh38.p14 and mouse112_GRCm39)}
\format{
data table
}
\usage{
data(gene_description)
}
\description{
gene_description: table containing gene description for human and mouse genes (human112_GRCh38.p14 and mouse112_GRCm39)
}
