% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_aberrant_celltypes.R
\name{list_aberrant_celltypes}
\alias{list_aberrant_celltypes}
\title{List aberrant cell types available in the Cell Marker Accordion disease database}
\usage{
list_aberrant_celltypes(
  species = c("Human", "Mouse"),
  disease = NULL,
  tissue = NULL
)
}
\arguments{
\item{species}{Character string or character string vector specifying the
species for which to extract the associate list of available cell types.
Currently, either “Human” and/or “Mouse” are supported. Default is
c("Mouse",“Human”).}

\item{disease}{Character string or character string vector specifying
diseases to consider. If NULL, information from all diseases are considered.
Default is NULL.}

\item{tissue}{Character string or character string vector specifying the tissue
for which to extract the associate list of available cell types.
If NULL, information from all tissues are retrieved.}
}
\value{
List of aberrant cell types available in the Cell Marker Accordion
disease database
}
\description{
List aberrant cell types available in the Cell Marker Accordion disease database
}
