% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_celltypes.R
\name{list_celltypes}
\alias{list_celltypes}
\title{List cell types available in the Cell Marker Accordion database}
\usage{
list_celltypes(species = c("Human", "Mouse"), tissue = NULL)
}
\arguments{
\item{species}{Character string or character string vector specifying the
species for which to extract the associate list of available cell types.
Currently, either “Human” and/or “Mouse” are supported. Default is
c("Mouse",“Human”).
@param tissue Character string or character string vector specifying the tissue
for which to extract the associate list of available cell types.
If NULL, information from all tissues are retrieved.}
}
\value{
List of cell types available in the Cell Marker Accordion database
}
\description{
List cell types available in the Cell Marker Accordion database
}
