% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_disease_tissues.R
\name{list_disease_tissues}
\alias{list_disease_tissues}
\title{List tissues available in the Cell Marker Accordion disease database}
\usage{
list_disease_tissues(
  species = c("Human", "Mouse"),
  disease = NULL,
  aberrant_celltype = NULL
)
}
\arguments{
\item{species}{Character string or character string vector specifying the
species for which to extract the associate list of available tissues.
Currently, either “Human” and/or “Mouse” are supported. Default is
c("Mouse",“Human”).}

\item{disease}{Character string or character string vector specifying
diseases to consider. If NULL, information from all diseases are considered.
Default is NULL.}

\item{celltype}{Character string or character string vector specifying the
celltype for which to extract the associate list of available tissues.
If NULL, information from all cell types are retrieved.}
}
\value{
List of tissues available in the Cell Marker Accordion disease
database
}
\description{
List tissues available in the Cell Marker Accordion disease database
}
