% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_tissues.R
\name{list_tissues}
\alias{list_tissues}
\title{List tissues available in the Cell Marker Accordion database}
\usage{
list_tissues(species = c("Human", "Mouse"), celltype = NULL)
}
\arguments{
\item{species}{Character string or character string vector specifying the
species for which to extract the associate list of available tissues.
Currently, either “Human” and/or “Mouse” are supported. Default is
c("Mouse",“Human”).}

\item{celltype}{Character string or character string vector specifying the
celltype for which to extract the associate list of available tissues.
If NULL, information from all cell types are retrieved.}
}
\value{
List of tissues available in the Cell Marker Accordion database
}
\description{
List tissues available in the Cell Marker Accordion database
}
