% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_database_integration.R
\name{marker_database_integration}
\alias{marker_database_integration}
\title{Create an augmented database by integrated custom set of markers genes with
the Accordion database}
\usage{
marker_database_integration(
  marker_table,
  database = "healthy",
  species_column = "species",
  disease_column = "disease",
  tissue_column = "tissue",
  celltype_column = "cell_type",
  marker_column = "marker",
  marker_type_column = "marker_type",
  resource_column = "resource"
)
}
\arguments{
\item{marker_table}{Data table or data frame containing cell type
markers. The table needs to have at least two columns, the
\code{celltype_column},  which specifies cell types, and the
\code{marker_column}, which specifies the corresponding markers on each row.
Columns indicating the marker type (either positive or negative), species, condition,
tissue and resource can be optionally included.}

\item{database}{String characters specifying the accordion database to use for
integration. Either “healthy” or “disease” are supported, where "healthy" refers
to the physiological accordion database, while "disease" refers to the disease
accordion database. Default is "healthy"}

\item{species_column}{String characters specifying the name of the
\code{marker_table} column containing species.
Default is “species”.}

\item{disease_column}{String characters specifying the name of the
\code{marker_table} column containing diseases.
Default is “disease”.}

\item{tissue_column}{String characters specifying the name of the
\code{marker_table} column containing tissues.
Default is “tissue”.}

\item{celltype_column}{String characters specifying the name of the
\code{marker_table} column containing cell types.
Default is “cell_type”.}

\item{marker_column}{String characters specifying the name of the
\code{marker_table} column containing markers.
Default is “marker”.}

\item{marker_type_column}{String characters specifying the name of the
\code{marker_table} column containing resources.
Default is “resource”.}
}
\value{
A data table
}
\description{
This function performs markers integration provided by the users and the
accordion database, either physiological ("healthy") or disease ("disease). It
takes in input a table of marker genes associated to cell types and return
the integrated database of custom and accordion markers. This new integrated
database can be used as input for the "accordion" or "disease_accordion" function
to provide the annotation.
}
