% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{marker_table}
\alias{marker_table}
\title{marker_table: Table of markers associated to retinal cell types}
\format{
A data table
}
\usage{
data(marker_table)
}
\description{
marker_table: Table of markers associated to retinal cell types
}
