% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{marker_table_pathway}
\alias{marker_table_pathway}
\title{marker_table_pathway: Table of markers associated to apoptosis}
\format{
A data table
}
\usage{
data(marker_table_pathway)
}
\description{
marker_table_pathway: Table of markers associated to apoptosis
}
