% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{mouse_data}
\alias{mouse_data}
\title{mouse_data: Mouse single-cell dataset}
\format{
A Seurat object
}
\usage{
data(mouse_data)
}
\description{
mouse_data: Mouse single-cell dataset
}
