% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{retinal_data}
\alias{retinal_data}
\title{retinal_data: Retina single-cell dataset}
\format{
A Seurat object
}
\usage{
data(retinal_data)
}
\description{
retinal_data: Retina single-cell dataset
}
