% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{uberon_onto}
\alias{uberon_onto}
\title{uberon_onto: Uberon anatomy ontology}
\format{
Uberon ontology
}
\usage{
data(uberon_onto)
}
\description{
uberon_onto: Uberon anatomy ontology
}
