% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accordion_plot.R
\name{accordion_plot}
\alias{accordion_plot}
\title{Interpreting annotation results}
\usage{
accordion_plot(
  data,
  info_to_plot = "accordion",
  resolution = "cluster",
  group_markers_by = "celltype_cluster"
)
}
\arguments{
\item{data}{A Seurat object or a list containing “detailed_annotation_info”,
from either accordion(), accordion_disease() or accordion_custom()
functions, run with include_detailed_annotation_info parameter set to TRUE.}

\item{info_to_plot}{Character string or character string vector specifying
the list from which extract the detailed annotation information, either
“accordion”, “accordion_disease” or “accordion_custom”, for which returns
the plot, either “accordion”, “accordion_disease” or “accordion_custom”.
Default is “accordion”.}

\item{resolution}{Character string or character string vector specifying the
annotation resolution for which provided the plots. Either “cluster” and/or
“cell” are supported. Default is “cluster”.}

\item{group_markers_by}{Character string or character string vector
specifying the classification of marker genes. It is possible to retrieve
top marker genes for each cell type identified with cluster
("celltype_cluster") or cell (“celltype_cell”) resolution; top  marker
genes per cluster ("cluster") or per cell ("cell") can be also obtained.
Additionally, by setting \code{group_markers_by}
to "score_cell", the \code{n_top_markers} marker genes only for
cells with a score greater than \code{top_cell_score_quantile_threshold} are
retrieved. Either "celltype_cluster", "celltype_cell",
"cluster", "cell" or "score_cell". Default is "celltype_cluster".}
}
\value{
A Seurat object or a list.
}
\description{
This function generates lollipop plots displaying the detailed annotation
results obtained with the \code{accordion_annotation},
\code{accordion_disease_annotation} and \code{accordion_custom_annotation}
functions.
}
\details{
top cell types (or pathways) and top markers It takes in input either a
Seurat object or a raw or normalized count matrix and a table of marker genes
associated to cell types or even to pathways and return in output the cell
types/pathways assignment and the detailed informations of the annotation
results (added to the Seurat object or as a list).

If a Seurat object was provided in input, the function returns the
Seurat object with a list of ggplot objects added to the "misc" slot in the
\code{info_to_plot} list. If a list was provided in input, the function
returns the same list with the addition of the ggplot objects.
}
