% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{uberon_onto}
\alias{uberon_onto}
\title{uberon_onto: Uberon anatomy ontology}
\format{
Uberon ontology

data table
}
\usage{
data(uberon_onto)

data(gene_description)
}
\description{
uberon_onto: Uberon anatomy ontology

gene_description: table containg gene description for human and mouse genes (human112_GRCh38.p14 and mouse112_GRCm39)
}
